feat(api): rename chatRepository to chatRedisRepository
