+++ api/src/llm/infrastructure/repositories/index.js
@@ -1,3 +1,3 @@
-export * as chatRepository from './chat-repository.js';
+export * as chatRedisRepository from './chat-redis-repository.js';
 export * as configurationRepository from './configuration-repository.js';
 export * as promptRepository from './prompt-repository.js';
