+++ api/tests/llm/acceptance/application/llm-preview-route_test.js
@@ -2,7 +2,7 @@ import { Readable } from 'node:stream';
 
 import { Chat } from '../../../../src/llm/domain/models/Chat.js';
 import { Configuration } from '../../../../src/llm/domain/models/Configuration.js';
-import { CHAT_STORAGE_PREFIX } from '../../../../src/llm/infrastructure/repositories/chat-repository.js';
+import { CHAT_STORAGE_PREFIX } from '../../../../src/llm/infrastructure/repositories/chat-redis-repository.js';
 import { featureToggles } from '../../../../src/shared/infrastructure/feature-toggles/index.js';
 import { temporaryStorage } from '../../../../src/shared/infrastructure/key-value-storages/index.js';
 import {
